requirements.txt
python-telegram-bot
google-generativeai
requests
python-dotenv